Docs: XML summaries added to IContentManager, ContentManagerPacked, and ContentManager.
Refactor: ContentManager and ContentManagerPacked moved to seperate files in same namespace.

Fix: ContentManagerPacked unload system for loaded resources added. (was missing)
