+++ ShapeEngine/Persistent/ContentManager.cs
@@ -1,0 +1,160 @@
+using Raylib_cs;
+
+namespace ShapeEngine.Persistent;
+
+/// <summary>
+/// Provides a simple class to load content and automatically unload all loaded content when close is called.
+/// </summary>
+public sealed class ContentManager : IContentManager
+{
+
+    private readonly List<Shader> shadersToUnload = [];
+    private readonly List<Texture2D> texturesToUnload = [];
+    private readonly List<Image> imagesToUnload = [];
+    private readonly List<Font> fontsToUnload = [];
+    private readonly List<Sound> soundsToUnload = [];
+    private readonly List<Music> musicToUnload = [];
+    private readonly List<Wave> wavesToUnload = [];
+
+    /// <summary>
+    /// Creates a new content manager.
+    /// </summary>
+    public ContentManager() { }
+
+    /// <summary>
+    /// Unloads all loaded content.
+    /// </summary>
+    public void Close()
+    {
+        foreach (var item in shadersToUnload)
+        {
+            Raylib.UnloadShader(item);
+        }
+
+        foreach (var item in texturesToUnload)
+        {
+            Raylib.UnloadTexture(item);
+        }
+        foreach (var item in imagesToUnload)
+        {
+            Raylib.UnloadImage(item);
+        }
+        foreach (var item in fontsToUnload)
+        {
+            Raylib.UnloadFont(item);
+        }
+        foreach (var item in wavesToUnload)
+        {
+            Raylib.UnloadWave(item);
+        }
+        foreach (var item in soundsToUnload)
+        {
+            Raylib.UnloadSound(item);
+        }
+        foreach (var item in musicToUnload)
+        {
+            Raylib.UnloadMusicStream(item);
+        }
+    }
+
+    /// <summary>
+    /// Loads a texture from the given file path.
+    /// </summary>
+    /// <param name="filePath">The path to the texture file.</param>
+    /// <returns>The loaded texture.</returns>
+    public Texture2D LoadTexture(string filePath)
+    {
+        var t = ContentLoader.LoadTexture(filePath);
+        texturesToUnload.Add(t);
+        return t;
+    }
+    /// <summary>
+    /// Loads an image from the given file path.
+    /// </summary>
+    /// <param name="filePath">The path to the image file.</param>
+    /// <returns>The loaded image.</returns>
+    public Image LoadImage(string filePath)
+    {
+        var i = ContentLoader.LoadImage(filePath);
+        imagesToUnload.Add(i);
+        return i;
+    }
+    /// <summary>
+    /// Loads a font from the given file path.
+    /// </summary>
+    /// <param name="filePath">The path to the font file.</param>
+    /// <param name="fontSize">The size of the font.</param>
+    /// <returns>The loaded font.</returns>
+    public Font LoadFont(string filePath, int fontSize = 100)
+    {
+        var f = ContentLoader.LoadFont(filePath, fontSize);
+        fontsToUnload.Add(f);
+        return f;
+    }
+    /// <summary>
+    /// Loads a wave from the given file path.
+    /// </summary>
+    /// <param name="filePath">The path to the wave file.</param>
+    /// <returns>The loaded wave.</returns>
+    public Wave LoadWave(string filePath)
+    {
+        var w = ContentLoader.LoadWave(filePath);
+        wavesToUnload.Add(w);
+        return w;
+    }
+    /// <summary>
+    /// Loads a sound from the given file path.
+    /// </summary>
+    /// <param name="filePath">The path to the sound file.</param>
+    /// <returns>The loaded sound.</returns>
+    public Sound LoadSound(string filePath)
+    {
+        var s = ContentLoader.LoadSound(filePath);
+        soundsToUnload.Add(s);
+        return s;
+
+    }
+    /// <summary>
+    /// Loads music from the given file path.
+    /// </summary>
+    /// <param name="filePath">The path to the music file.</param>
+    /// <returns>The loaded music.</returns>
+    public Music LoadMusic(string filePath)
+    {
+        var m = ContentLoader.LoadMusicStream(filePath);
+        musicToUnload.Add(m);
+        return m;
+    }
+    /// <summary>
+    /// Loads a fragment shader from the given file path.
+    /// </summary>
+    /// <param name="filePath">The path to the fragment shader file.</param>
+    /// <returns>The loaded shader.</returns>
+    public Shader LoadFragmentShader(string filePath)
+    {
+        var fs = ContentLoader.LoadFragmentShader(filePath);
+        shadersToUnload.Add(fs);
+        return fs;
+    }
+    /// <summary>
+    /// Loads a vertex shader from the given file path.
+    /// </summary>
+    /// <param name="filePath">The path to the vertex shader file.</param>
+    /// <returns>The loaded shader.</returns>
+    public Shader LoadVertexShader(string filePath)
+    {
+        var vs = ContentLoader.LoadVertexShader(filePath);
+        shadersToUnload.Add(vs);
+        return vs;
+    }
+    /// <summary>
+    /// Loads a json string from the given file path.
+    /// </summary>
+    /// <param name="filePath">The path to the json file.</param>
+    /// <returns>The loaded json string.</returns>
+    public string LoadJson(string filePath)
+    {
+        return ContentLoader.LoadJson(filePath);
+    }
+
+}
